refactor: move REST handlers and services
