+++ api/cmd/serverd/main.go
@@ -41,11 +41,11 @@ import (
 	_ "github.com/erwin-lovecraft/aegismiles/docs"
 	"github.com/erwin-lovecraft/aegismiles/internal/adapters/gateway/auth0"
 	"github.com/erwin-lovecraft/aegismiles/internal/adapters/gateway/sessionm"
+	"github.com/erwin-lovecraft/aegismiles/internal/adapters/handler/rest/middleware"
+	v1 "github.com/erwin-lovecraft/aegismiles/internal/adapters/handler/rest/v1"
+	v2 "github.com/erwin-lovecraft/aegismiles/internal/adapters/handler/rest/v2"
 	"github.com/erwin-lovecraft/aegismiles/internal/config"
 	"github.com/erwin-lovecraft/aegismiles/internal/constants"
-	"github.com/erwin-lovecraft/aegismiles/internal/controller/rest/middleware"
-	v1 "github.com/erwin-lovecraft/aegismiles/internal/controller/rest/v1"
-	v2 "github.com/erwin-lovecraft/aegismiles/internal/controller/rest/v2"
 	"github.com/erwin-lovecraft/aegismiles/internal/core/service/customer"
 	"github.com/erwin-lovecraft/aegismiles/internal/core/service/mileage"
 	"github.com/erwin-lovecraft/aegismiles/internal/pkg/generator"
