+++ api/internal/adapters/handler/rest/middleware/has_role.go
