+++ api/internal/adapters/handler/rest/v1/controller.go
@@ -1,4 +1,4 @@
-package v2
+package rest
 
 import (
 	"net/http"
