+++ api/internal/adapters/handler/rest/v2/controller.go
@@ -1,4 +1,4 @@
-package v1
+package rest
 
 import (
 	"net/http"
